requirements.txt
Flask>=2.0.0
PyPDF2>=3.0.0
google-generativeai==0.8.3
python-dotenv>=0.19.0
